+++ packages/patrol_cli/lib/src/commands/build_android.dart
@@ -138,6 +138,7 @@ class BuildAndroidCommand extends PatrolCommand {
 
     final displayLabel = boolArg('label');
     final uninstall = boolArg('uninstall');
+    final noTreeShakeIcons = boolArg('no-tree-shake-icons');
 
     final customDartDefines = {
       ..._dartDefinesReader.fromFile(),
@@ -183,6 +184,7 @@ class BuildAndroidCommand extends PatrolCommand {
       dartDefineFromFilePaths: dartDefineFromFilePaths,
       buildName: buildName,
       buildNumber: buildNumber,
+      noTreeShakeIcons: noTreeShakeIcons,
     );
 
     final androidOpts = AndroidAppOptions(