+++ packages/patrol_cli/lib/src/commands/build_ios.dart
@@ -180,6 +180,7 @@ class BuildIOSCommand extends PatrolCommand {
       dartDefineFromFilePaths: dartDefineFromFilePaths,
       buildName: buildName,
       buildNumber: buildNumber,
+      noTreeShakeIcons: boolArg('no-tree-shake-icons'),
     );
 
     final iosOpts = IOSAppOptions(